适配template至1.8.0

+++ user/template/protocol.h
@@ -211,10 +211,10 @@ typedef struct {
 typedef struct {
     union {
         struct {
-            float data1;
-            float data2;
-            float data3;
-            float data4;
+            int32_t data1;
+            int32_t data2;
+            int32_t data3;
+            int32_t data4;
         };
         struct {
             uint8_t data[16];
@@ -228,10 +228,10 @@ typedef struct {
             uint16_t data_cmd_id;
             uint16_t send_id;
             uint16_t receiver_id;
-            float    data1;
-            float    data2;
-            float    data3;
-            float    data4;
+            int32_t  data1;
+            int32_t  data2;
+            int32_t  data3;
+            int32_t  data4;
         };
         struct {
             uint8_t data[22];
